refactor(ai): move ChatConfig and StartChatAbortPollingArgs to types/chat
