+++ apps/dashboard/src/app/api/organizations/[organizationId]/chat/route.ts
@@ -10,12 +10,12 @@ import {
   replaceChatHistory,
   setActiveChatStream,
 } from "@notra/ai/chat/history";
-import { buildChatFinishMetadata } from "@notra/ai/utils/chat";
-import { standaloneChatRequestSchema } from "@notra/ai/schemas/chat";
-import type { ChatUsageSnapshot } from "@notra/ai/types/chat";
 import { orchestrateStandaloneChat } from "@notra/ai/orchestration/orchestrate-standalone";
+import { standaloneChatRequestSchema } from "@notra/ai/schemas/chat";
 import type { StandaloneChatContextItem } from "@notra/ai/schemas/standalone-chat";
+import type { ChatUsageSnapshot } from "@notra/ai/types/chat";
 import type { ValidatedIntegration } from "@notra/ai/types/orchestration";
+import { buildChatFinishMetadata } from "@notra/ai/utils/chat";
 import type { UIMessage } from "ai";
 import type { CheckResponse } from "autumn-js";
 import { nanoid } from "nanoid";
