+++ apps/dashboard/src/app/api/workflows/chat/route.ts
@@ -6,14 +6,14 @@ import {
   loadChatHistory,
   replaceChatHistory,
 } from "@notra/ai/chat/history";
-import { buildChatFinishMetadata } from "@notra/ai/utils/chat";
+import { orchestrateStandaloneChat } from "@notra/ai/orchestration/orchestrate-standalone";
 import { chatWorkflowPayloadSchema } from "@notra/ai/schemas/chat";
 import type {
   ChatUsageSnapshot,
   ChatWorkflowPayload,
 } from "@notra/ai/types/chat";
-import { orchestrateStandaloneChat } from "@notra/ai/orchestration/orchestrate-standalone";
 import type { StandaloneChatContextItem } from "@notra/ai/types/standalone-chat";
+import { buildChatFinishMetadata } from "@notra/ai/utils/chat";
 import { serve } from "@upstash/workflow/nextjs";
 import type { UIMessageChunk } from "ai";
 import { nanoid } from "nanoid";
